+++ crates/squawk_fmt/src/lib.rs
@@ -1,3 +1,4 @@
 mod fmt;
+mod ir;
 
-pub use fmt::fmt;
+pub use fmt::{fmt, fmt_ir};